backend/requirements.txt
fastapi
uvicorn
python-multipart
mysql-connector-python
bcrypt
pyjwt
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-experimental
faiss-cpu
pypdf
python-docx
google-generativeai
sentence-transformers